preperties edit

+++ src/main/java/com/example/deligo/order/repository/OrderRepository.java
@@ -4,6 +4,8 @@ import com.example.deligo.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
